requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Django==3.0.5
djangorestframework==3.12.1
djongo==1.3.3
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.1
httplib2==0.18.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
